+++ packages/eas-cli/src/project/ios/versions.ts
@@ -1,6 +1,6 @@
 import assert from 'assert';
 
-export const INVALID_BUILD_NUMBER_ERROR_MESSAGE = `Invalid value, buildNumber needs to consist only of up to 3 dot-separated positive integers.`;
+export const BUILD_NUMBER_REQUIREMENTS = `buildNumber needs to consist only of up to 3 dot-separated positive integers`;
 
 export function isValidBuildNumber(buildNumber: string): boolean {
   return !!buildNumber.match(/^\d+(\.\d+)?(\.\d+)?$/);